Shrikant Kosariya,B Kavya,Sagar Sahasi
The User can query any college-related activities through the system. The user does not have to personally go to the college for inquiry.
Students just have to query through the bot which is used for chatting.
A chatbot is artificial intelligence (AI) software that can simulate a conversation (or a chat) with a user in natural language through messaging applications, websites, mobile apps or through the telephone.